Move data type definitions to constants file
