+++ frontend/src/components/eMIB/constants.js
@@ -1,2 +1,10 @@
+// Display
 export const HEADER_HEIGHT = 130;
 export const FOOTER_HEIGHT = 72;
+
+// Data structure
+export const EMAIL_TYPE = {
+  reply: "reply",
+  replyAll: "reply all",
+  forward: "forward"
+};
